requirements.txt
Flask==2.0.1
#gunicorn==19.4.5
gunicorn==20.0.4
python-decouple==3.4
Flask-RESTful==0.3.9
# celery==4.4.0
# redis==3.3.7
selenium==3.141.0
ordered-set==4.1.0
bs4==0.0.1